fix: preserve cli help and provider split
